requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
django-environ==0.9.0
django-extensions==3.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.4
mysqlclient==2.1.1
Pillow==9.3.0
PyJWT==2.5.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
